requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
pydantic==2.13.3
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-huggingface==0.1.0
chromadb==0.5.5
sentence-transformers==3.1.0
python-dotenv==1.0.1
